FCA #12 - Removing GoogleAnalytics in favor of CookiePolicy
